code style fixed

+++ tasks/image_processor/Filters/blur_filter.cpp
@@ -11,7 +11,7 @@ struct DoublePixel {
 };
 
 const DoublePixel GetKernelMatrixValueHorizontal(int32_t y, int32_t x, ColorMatrix& matrix, double sigma,
-                                                 double* e_array) {
+                                                 long double* e_array) {
     int32_t dst = static_cast<int32_t>(BlurFilter::FILTER_SIGMA_COEFFICIENT * sigma);
 
     DoublePixel pixel;
@@ -19,7 +19,7 @@ const DoublePixel GetKernelMatrixValueHorizontal(int32_t y, int32_t x, ColorMatr
     for (int32_t x_coordinate = x - dst; x_coordinate <= x + dst; ++x_coordinate) {
         int32_t actual_x = std::max(0, std::min(static_cast<int32_t>(matrix.GetWidth()) - 1, x_coordinate));
         Pixel window_pixel = matrix.GetPixel(y, actual_x);
-        double divisor = e_array[abs(x - x_coordinate)];
+        long double divisor = e_array[abs(x - x_coordinate)];
         pixel.R += static_cast<long double>(window_pixel.R) / divisor;
         pixel.G += static_cast<long double>(window_pixel.G) / divisor;
         pixel.B += static_cast<long double>(window_pixel.B) / divisor;
@@ -28,8 +28,8 @@ const DoublePixel GetKernelMatrixValueHorizontal(int32_t y, int32_t x, ColorMatr
     return pixel;
 }
 
-const DoublePixel GetKernelMatrixValueVertical(int32_t y, int32_t x, double** matrix, double sigma, int32_t height,
-                                               double* e_array) {
+const DoublePixel GetKernelMatrixValueVertical(int32_t y, int32_t x, long double** matrix, double sigma, int32_t height,
+                                               long double* e_array) {
     int32_t dst = static_cast<int32_t>(BlurFilter::FILTER_SIGMA_COEFFICIENT * sigma);
 
     DoublePixel pixel;
@@ -54,7 +54,7 @@ Pixel CastDoublePixel(DoublePixel pixel) {
 
 void BlurFilter::Apply(BMP& bmp_stream) {
     double sigma = std::stod(setting_.GetFilterParameter(0));
-    double* e_array = new double[static_cast<int32_t>(FILTER_SIGMA_COEFFICIENT * sigma) + 1];
+    long double* e_array = new long double[static_cast<int32_t>(FILTER_SIGMA_COEFFICIENT * sigma) + 1];
 
     for (int32_t dst = 0; dst <= static_cast<int32_t>(FILTER_SIGMA_COEFFICIENT * sigma); ++dst) {
         double power = static_cast<double>(dst * dst) / 2 / sigma / sigma;
@@ -64,12 +64,12 @@ void BlurFilter::Apply(BMP& bmp_stream) {
     int32_t height = bmp_stream.GetDibHeader().height;
     int32_t width = bmp_stream.GetDibHeader().width;
 
-    double** horizontal_matrix = new double*[height];
+    long double** horizontal_matrix = new long double*[height];
 
     double multiplier = 1.0 / sigma / sigma / 2 / PI;
 
     for (int32_t y_coordinate = 0; y_coordinate < height; ++y_coordinate) {
-        horizontal_matrix[y_coordinate] = new double[3 * width];
+        horizontal_matrix[y_coordinate] = new long double[3 * width];
         for (int32_t x_coordinate = 0; x_coordinate < width; ++x_coordinate) {
             DoublePixel pixel =
                 GetKernelMatrixValueHorizontal(y_coordinate, x_coordinate, bmp_stream.GetPixelArray(), sigma, e_array);
